Refactor particle pool storage to Chunk level and optimize propagation
- Moved `particlePool` field from `CardPlayer` and `Structure` to `Chunk`.
- Updated `Actor` interface to delegate `particlePool()` calls to the parent chunk.
- Updated `World.particlePool` to propagate the pool to all loaded chunks.
- Ensured `particlePool` is correctly set during `Zone` generation and `reindex`.
- Verified all tests pass.
