+++ nomad-realms/src/main/java/nomadrealms/context/game/actor/types/structure/Structure.java
@@ -25,8 +25,6 @@ import nomadrealms.render.particle.ParticlePool;
 
 public abstract class Structure implements Actor {
 
-	private transient ParticlePool particlePool = new NullParticlePool();
-
 	private TileCoordinate tileCoord;
 	private transient Tile tile;
 	private Inventory inventory;
@@ -130,7 +128,6 @@ public abstract class Structure implements Actor {
 		if (inventory != null) {
 			inventory.reindex(this);
 		}
-		particlePool(world.particlePool());
 	}
 
 	@Override
@@ -141,16 +138,6 @@ public abstract class Structure implements Actor {
 	@Override
 	public Status status() {
 		return status;
-	}
-
-	@Override
-	public void particlePool(ParticlePool particlePool) {
-		this.particlePool = particlePool;
-	}
-
-	@Override
-	public ParticlePool particlePool() {
-		return particlePool;
 	}
 
 	public abstract StructureType structureType();
