+++ nomad-realms/src/main/java/nomadrealms/context/game/world/World.java
@@ -183,7 +183,6 @@ public class World {
 			}
 		}
 		for (Actor actor : actorsToUpdate) {
-			actor.particlePool(particlePool());
 			if (actor.isDestroyed()) {
 				if (actor.tile() != null) {
 					spawnDeathParticles(actor);
@@ -252,9 +251,6 @@ public class World {
 	}
 
 	public void addActor(Actor actor, boolean forced) {
-		if (state != null) {
-			actor.particlePool(state.particlePool);
-		}
 		if (forced) {
 			actor.tile().clearActor();
 		}
@@ -330,11 +326,18 @@ public class World {
 	}
 
 	public void particlePool(ParticlePool particlePool) {
-		if (nomad != null && nomad.tile() != null) {
-			for (Chunk chunk : nomad.tile().chunk().getSurroundingChunks()) {
-				if (chunk != null) {
-					for (Actor actor : chunk.actors()) {
-						actor.particlePool(particlePool);
+		for (Region region : map.regions()) {
+			for (Zone[] zoneRow : region.zones()) {
+				for (Zone zone : zoneRow) {
+					if (zone == null) {
+						continue;
+					}
+					for (Chunk[] chunkRow : zone.chunks()) {
+						for (Chunk chunk : chunkRow) {
+							if (chunk != null) {
+								chunk.particlePool(particlePool);
+							}
+						}
 					}
 				}
 			}
