+++ nomad-realms/src/main/java/nomadrealms/context/game/world/map/area/Zone.java
@@ -75,10 +75,8 @@ public class Zone {
 		if (world.state() != null) {
 			for (Chunk[] row : chunks) {
 				for (Chunk chunk : row) {
-					for (Tile tile : chunk.tiles()) {
-						if (tile.actor() != null) {
-							tile.actor().particlePool(world.state().particlePool);
-						}
+					if (chunk != null) {
+						chunk.particlePool(world.state().particlePool);
 					}
 				}
 			}
@@ -178,6 +176,9 @@ public class Zone {
 			for (Chunk chunk : chunkRow) {
 				if (chunk != null) {
 					chunk.reindex(this);
+					if (world.state() != null) {
+						chunk.particlePool(world.state().particlePool);
+					}
 				}
 			}
 		}
